imperfect_scrobbles/requirements.txt
sentence-transformers>=2.7.0
pandas>=2.0.0,<2.1.0
pyyaml>=6.0.1
numpy>=1.24.0,<1.26.0
scikit-learn>=1.3.0,<1.4.0
matplotlib>=3.7.0,<3.8.0
pykakasi>=2.2.1
cutlet>=0.3.0
pypinyin>=0.50.0
korean-romanizer>=0.0.7
unidecode>=1.3.7
